factor out versioning
